requirements.txt
PySide6>=6.6,<7.0
qt-material>=2.14
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.126.0
markdown>=3.6
